generators/wikipedia/requirements.txt
lxml==4.3.2
mwclient==0.9.3
Pillow==5.4.1
python-pptx==0.6.18
python-slugify==3.0.0
requests==2.21.0
nltk==3.4.3
python-resize-image==1.1.19
urllib3==1.24.1
